backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.0
pandas==2.2.3
faiss-cpu
requests
langchain
langchain_community
langchain_core
langchain-ollama
ollama